fix: Astro proxy forwards bodies, XFF, and Set-Cookie correctly
- Add shared web/src/lib/proxy.ts with createProxy():

  * send duplex: half for POST/PUT/PATCH bodies

  * preserve X-Forwarded-For instead of deleting it

  * forward each Set-Cookie individually via getSetCookie()

  * redirect: manual so OAuth 302s reach the client

- Refactor /api/[...path].ts and /gh/[...path].ts to use createProxy.
