+++ .pr-body.tmp.md
@@ -1,0 +1,47 @@
+Step 9: Deploy to Hetzner VPS with Kamal + Astro skeleton + OpenAPI
+
+## Summary
+
+Completes the Step 9 plan in `.plans/STEP9-KAMALPLAN.md` with review fixes applied:
+
+1. **Dockerize the API** — multi-stage `api/Dockerfile` building `androidskills-api-all.jar` with Eclipse Temurin.
+2. **Kamal 2.x scaffold** — `deploy/kamal.yml` is the **default, deployable fallback** (kamal-proxy → web → Astro proxies `/api/*` and `/gh/*` to Ktor). `deploy/kamal.path-prefix.yml` is the aspirational config for direct kamal-proxy path-prefix routing once confirmed.
+3. **Astro skeleton** — `/web` Astro 5 + `@astrojs/node` standalone SSR project, placeholder pages, and proxy routes for `/api/*` and `/gh/*`.
+4. **Litestream backup/restore** — `api/litestream.yml` + `api/entrypoint.sh` restores the SQLite DB from R2 on boot only if a replica exists (first deploy boots fresh), then runs Ktor under Litestream replication.
+5. **OpenAPI contract** — `api/src/main/resources/openapi.yaml` served at `GET /api/openapi.yaml`; covers public/auth/contributor routes, excludes admin paths, and is validated against real responses.
+6. **GitHub Actions** — `ci.yml` runs `./gradlew test` and `npm ci && npm run build` on PRs/pushes to `develop` and `main`; `deploy.yml` deploys on `main` push and `workflow_dispatch`.
+7. **Step 8 follow-ups** — stricter `PUBLIC_BASE_URL` scheme/host/path validation, `TRUSTED_PROXY_COUNT` docs.
+
+## Verified
+
+- `./gradlew test` in `api/` passes (262+ tests).
+- `npm ci && npm run build` in `web/` succeeds and produces `dist/server/entry.mjs`.
+- OpenAPI YAML parses as valid YAML and `OpenApiContractTest` asserts real response shapes against the spec.
+
+## Review fixes since first pass
+
+- `api/entrypoint.sh` now uses `litestream restore -if-replica-exists` so first deploy boots fresh.
+- Default Kamal topology switched to fallback (Astro proxies `/api/*` and `/gh/*`), fixing the missing `/gh/webhooks` route in the primary config.
+- `deploy.yml` no longer runs `kamal lock release`; `SEED_DEMO` removed from production secrets.
+- Added `.github/workflows/ci.yml` as a green gate.
+- Admin routes removed from the public OpenAPI spec; test now asserts real response shapes.
+- Containers run as non-root UID/GID 1000; host volume must be owned accordingly.
+- `api/litestream.yml` path uses `${DATA_DIR:-/data}`.
+
+## Not verified locally (flagged for review)
+
+- Docker build could not be run (Docker Desktop daemon not available).
+- Kamal is not installed locally; `kamal.yml` and `kamal.path-prefix.yml` syntax is untested and includes `REVIEW` comments for the exact Kamal-version options (no-overlap deploy, stop-grace, volume syntax).
+- The GitHub Actions workflows are scaffolds; they will need the full secret set before the first run.
+
+## Files to review
+
+- `api/Dockerfile`, `api/entrypoint.sh`, `api/litestream.yml`
+- `deploy/kamal.yml`, `deploy/kamal.path-prefix.yml`
+- `web/Dockerfile`, `web/astro.config.mjs`, `web/src/pages/api/[...path].ts`, `web/src/pages/gh/[...path].ts`
+- `api/src/main/resources/openapi.yaml`
+- `api/src/main/kotlin/dev/androidskills/Application.kt` (`/api/openapi.yaml` route)
+- `api/src/test/kotlin/dev/androidskills/api/OpenApiContractTest.kt`
+- `api/src/main/kotlin/dev/androidskills/AppConfig.kt` (validation)
+- `.github/workflows/ci.yml`, `.github/workflows/deploy.yml`
+- `deploy/README.md`, `api/README.md`
